refactor: update imports and configuration for improved structure and readability
- Changed the export method in `tailwind.config.js` to use ES6 syntax for consistency with modern JavaScript practices.
- Removed unused imports and adjusted import paths in various components to utilize the new `useNotification` hook and `notificationTypes` constant, enhancing code clarity and maintainability.
- Updated the `Home` component to correctly reference the `user` object instead of `userData`, ensuring accurate user information display.
- Minor adjustments made across multiple files to streamline the codebase and improve overall organization.
- This refactor aims to enhance code readability and maintainability while adhering to best practices in the codebase.
