+++ inventary/src/components/NotificationSystem/Notification.jsx
@@ -5,7 +5,7 @@ import CheckCircleIcon from "@mui/icons-material/CheckCircle";
 import ErrorIcon from "@mui/icons-material/Error";
 import InfoIcon from "@mui/icons-material/Info";
 import WarningIcon from "@mui/icons-material/Warning";
-import { NOTIFICATION_TYPES } from "../../contexts/NotificationContext";
+import { NOTIFICATION_TYPES } from "../../constants/notificationTypes";
 
 // Get icon based on notification type
 const getIcon = (type) => {
