+++ inventary/src/constants/notificationTypes.js
@@ -1,0 +1,8 @@
+// Notification types
+
+export const NOTIFICATION_TYPES = {
+  SUCCESS: "success",
+  ERROR: "error",
+  WARNING: "warning",
+  INFO: "info",
+};
