+++ inventary/src/pages/MasterProduct/index.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo, useRef, useCallback } from "react";
+import React, { useState, useEffect, useMemo, useRef } from "react";
 import {
   CircularProgress,
   Typography,
